+++ lib/ckb/api.rb
@@ -16,7 +16,7 @@ module CKB
       @rpc = CKB::RPC.new(host: host)
       if mode == MODE::TESTNET
         # Testnet system script code_hash
-        expected_code_hash = "0x9e3b3557f11b2b3532ce352bfe8017e9fd11d154c4c7f9b7aaaa1e621b539a08"
+        expected_code_hash = "0xf1951123466e4479842387a66fabfd6b65fc87fd84ae8e6cd3053edb27fff2fd"
         # For testnet chain, we can assume the second cell of the first transaction
         # in the genesis block contains default lock script we can use here.
         system_cell_transaction = genesis_block.transactions.first
@@ -27,7 +27,7 @@ module CKB
           )
         )
         cell_data = CKB::Utils.hex_to_bin(system_cell_transaction.outputs[1].data)
-        code_hash = CKB::Utils.bin_to_hex(CKB::Blake2b.digest(cell_data))
+        code_hash = CKB::Blake2b.hexdigest(cell_data)
 
         raise "System script code_hash error!" unless code_hash == expected_code_hash